Backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.7
websockets>=12.0
tensorflow>=2.15.0
numpy>=1.24.0
Pillow>=10.0.0
